Refactor type representation and type handling in Malgo
- Moved `HasType` instances from `Malgo.Syntax` to `Malgo.Infer.TypeRep` to centralize type handling.
- Updated `Typed` data type and its instances for better type management.
- Modified import statements to reflect the new structure and removed redundant instances.
- Adjusted the dependency graph generation command in `mise.toml` to exclude `Malgo.SExpr`.
- Marked the completion of moving `HasType` instances in the todo list.
