+++ dependencies.dot
@@ -13,6 +13,8 @@ digraph G {
     "Malgo.Driver" -> "Malgo.Sequent.Eval";
     "Malgo.Driver" -> "Malgo.Sequent.ToCore";
     "Malgo.Driver" -> "Malgo.Sequent.ToFun";
+    "Malgo.Driver" -> "Malgo.Syntax";
+    "Malgo.Driver" -> "Malgo.Syntax.Extension";
     "Malgo.Infer" -> "Malgo.Infer.Error";
     "Malgo.Infer" -> "Malgo.Infer.Kind";
     "Malgo.Infer" -> "Malgo.Infer.Pass";
@@ -30,11 +32,17 @@ digraph G {
     "Malgo.Infer.Pass" -> "Malgo.Module";
     "Malgo.Infer.Pass" -> "Malgo.Pass";
     "Malgo.Infer.Pass" -> "Malgo.Rename";
+    "Malgo.Infer.Pass" -> "Malgo.Syntax";
+    "Malgo.Infer.Pass" -> "Malgo.Syntax";
+    "Malgo.Infer.Pass" -> "Malgo.Syntax.Extension";
     "Malgo.Infer.TcEnv" -> "Malgo.Infer.Kind";
     "Malgo.Infer.TcEnv" -> "Malgo.Infer.TypeRep";
     "Malgo.Infer.TcEnv" -> "Malgo.Interface";
     "Malgo.Infer.TcEnv" -> "Malgo.Module";
     "Malgo.Infer.TcEnv" -> "Malgo.Rename";
+    "Malgo.Infer.TcEnv" -> "Malgo.Syntax.Extension";
+    "Malgo.Infer.TypeRep" -> "Malgo.Syntax";
+    "Malgo.Infer.TypeRep" -> "Malgo.Syntax.Extension";
     "Malgo.Infer.Unify" -> "Malgo.Infer.Error";
     "Malgo.Infer.Unify" -> "Malgo.Infer.Kind";
     "Malgo.Infer.Unify" -> "Malgo.Infer.TcEnv";
@@ -43,14 +51,17 @@ digraph G {
     "Malgo.Interface" -> "Malgo.Infer.Kind";
     "Malgo.Interface" -> "Malgo.Infer.TypeRep";
     "Malgo.Interface" -> "Malgo.Module";
-    "Malgo.Module" -> "Malgo.SExpr";
-    "Malgo.Module" -> "Malgo.SExpr";
+    "Malgo.Interface" -> "Malgo.Syntax.Extension";
     "Malgo.Monad" -> "Malgo.Interface";
     "Malgo.Monad" -> "Malgo.Module";
     "Malgo.Parser" -> "Malgo.Module";
+    "Malgo.Parser" -> "Malgo.Syntax";
+    "Malgo.Parser" -> "Malgo.Syntax.Extension";
     "Malgo.Parser.Pass" -> "Malgo.Module";
     "Malgo.Parser.Pass" -> "Malgo.Parser";
     "Malgo.Parser.Pass" -> "Malgo.Pass";
+    "Malgo.Parser.Pass" -> "Malgo.Syntax";
+    "Malgo.Parser.Pass" -> "Malgo.Syntax.Extension";
     "Malgo.Refine" -> "Malgo.Refine.Pass";
     "Malgo.Refine" -> "Malgo.Refine.RefineEnv";
     "Malgo.Refine" -> "Malgo.Refine.Space";
@@ -58,9 +69,14 @@ digraph G {
     "Malgo.Refine.Pass" -> "Malgo.Pass";
     "Malgo.Refine.Pass" -> "Malgo.Refine.RefineEnv";
     "Malgo.Refine.Pass" -> "Malgo.Refine.Space";
+    "Malgo.Refine.Pass" -> "Malgo.Syntax";
+    "Malgo.Refine.Pass" -> "Malgo.Syntax";
+    "Malgo.Refine.Pass" -> "Malgo.Syntax.Extension";
     "Malgo.Refine.RefineEnv" -> "Malgo.Infer";
     "Malgo.Refine.Space" -> "Malgo.Infer";
     "Malgo.Refine.Space" -> "Malgo.Refine.RefineEnv";
+    "Malgo.Refine.Space" -> "Malgo.Syntax";
+    "Malgo.Refine.Space" -> "Malgo.Syntax.Extension";
     "Malgo.Rename" -> "Malgo.Rename.Pass";
     "Malgo.Rename" -> "Malgo.Rename.RnEnv";
     "Malgo.Rename" -> "Malgo.Rename.RnState";
@@ -69,10 +85,13 @@ digraph G {
     "Malgo.Rename.Pass" -> "Malgo.Pass";
     "Malgo.Rename.Pass" -> "Malgo.Rename.RnEnv";
     "Malgo.Rename.Pass" -> "Malgo.Rename.RnState";
+    "Malgo.Rename.Pass" -> "Malgo.Syntax";
+    "Malgo.Rename.Pass" -> "Malgo.Syntax.Extension";
     "Malgo.Rename.RnEnv" -> "Malgo.Module";
+    "Malgo.Rename.RnEnv" -> "Malgo.Syntax.Extension";
     "Malgo.Rename.RnState" -> "Malgo.Module";
+    "Malgo.Rename.RnState" -> "Malgo.Syntax.Extension";
     "Malgo.Sequent.Core" -> "Malgo.Module";
-    "Malgo.Sequent.Core" -> "Malgo.SExpr";
     "Malgo.Sequent.Core" -> "Malgo.Sequent.Fun";
     "Malgo.Sequent.Core.Flat" -> "Malgo.Module";
     "Malgo.Sequent.Core.Flat" -> "Malgo.Pass";
@@ -84,19 +103,22 @@ digraph G {
     "Malgo.Sequent.Core.Join" -> "Malgo.Sequent.Fun";
     "Malgo.Sequent.Eval" -> "Malgo.Module";
     "Malgo.Sequent.Eval" -> "Malgo.Pass";
-    "Malgo.Sequent.Eval" -> "Malgo.SExpr";
     "Malgo.Sequent.Eval" -> "Malgo.Sequent.Core";
     "Malgo.Sequent.Eval" -> "Malgo.Sequent.Fun";
     "Malgo.Sequent.Fun" -> "Malgo.Module";
-    "Malgo.Sequent.Fun" -> "Malgo.SExpr";
-    "Malgo.Sequent.Fun" -> "Malgo.SExpr";
     "Malgo.Sequent.ToCore" -> "Malgo.Module";
     "Malgo.Sequent.ToCore" -> "Malgo.Pass";
     "Malgo.Sequent.ToCore" -> "Malgo.Sequent.Core";
     "Malgo.Sequent.ToCore" -> "Malgo.Sequent.Core";
     "Malgo.Sequent.ToCore" -> "Malgo.Sequent.Fun";
-    "Malgo.Sequent.ToFun" -> "Malgo.Infer.TypeRep";
+    "Malgo.Sequent.ToFun" -> "Malgo.Infer";
+    "Malgo.Sequent.ToFun" -> "Malgo.Infer";
     "Malgo.Sequent.ToFun" -> "Malgo.Module";
     "Malgo.Sequent.ToFun" -> "Malgo.Pass";
     "Malgo.Sequent.ToFun" -> "Malgo.Sequent.Fun";
+    "Malgo.Sequent.ToFun" -> "Malgo.Syntax";
+    "Malgo.Sequent.ToFun" -> "Malgo.Syntax.Extension";
+    "Malgo.Syntax" -> "Malgo.Module";
+    "Malgo.Syntax" -> "Malgo.Syntax.Extension";
+    "Malgo.Syntax.Extension" -> "Malgo.Module";
 }
